# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Projects
...
  /projects/{id}/deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"

    post:
      operationId: triggerDeployment
      tags: [deployments]
      summary: Trigger a new deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentTriggerInput"
      responses:
        "201":
          description: Deployment queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /projects/{id}/deployments/{deploymentId}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details and logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /projects/{id}/deployments/{deploymentId}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deployments]
      summary: Rollback to this deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Rollback deployment queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  # Project actions
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 2
          maxLength: 80
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, plan, credits, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [hobby, pro, team]
        credits:
          type: integer
        createdAt:
          type: string

    # Project schemas
...
    Deployment:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [queued, building, deploying, running, failed, stopped, rolled_back]
        commitHash:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        buildLog:
          type: ["string", "null"]
        createdAt:
          type: string
        deployedAt:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
    DeploymentTriggerInput:
      type: object
      properties:
        commitHash:
          type: string
        commitMessage:
          type: string

    # Database schemas
...