# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}