# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
}